pyproject.toml
[project]
name = "modelbridge"
version = "0.1.0"
description = ""
authors = [
    {name = "Name",email = "mail@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "loguru (>=0.7.3,<0.8.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "sqlalchemy (>=2.0.38,<3.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "httpx (>=0.28.1,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
